Add separate test for counting_sort
